IMPALA-12508: Update query option "enabled_runtime_filter_types" comments
The query option "enabled_runtime_filter_types" currently allows
specifying a comma-separated list of runtime filter types (including
BLOOM, MIN_MAX and IN_LIST) to enable or setting it to "all" to use all
types. But the comment for this query option does not mention IN_LIST
and states that it only applies to Kudu, which is outdated, as it also
affects HDFS scans after IMPALA-10325.

Therefore, this patch update the relevant comments to avoid any
misleading caused by the outdated comments.

Change-Id: Iecff7c655e273712d851d9ce94ef5713e579aa72
Reviewed-on: http://gerrit.cloudera.org:8080/20627
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>