IMPALA-12529: Show logs for the time of acquiring table/db locks
Waiting for table locks could take long when there are lots of
concurrent catalog operations, e.g. DDL/DMLs or HMS events processing,
on the same table. It'd be helpful to show the time of acquiring table
locks in logs for better diagnostics.

We already have TRACE level logs for acquiring the table/db locks. This
changes them to WARN level when the duration exceeds a threshold
(100ms).

Ran a GVO without the threshold and found most of the duration are under
100ms. The majority is 0ms. So choose 100ms to just show abnormal
locking.

Change-Id: I99d0665a5ab815a81a5341f63aff1ac2f9a447d5
Reviewed-on: http://gerrit.cloudera.org:8080/20643
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>