IMPALA-11776: Use hive.metastore.table.owner for create iceberg table
IMPALA-11429 made the creation of an Iceberg table happen in 2 steps.
The first step creates the table, however, with wrong owner and the
second step is an ALTER TABLE to set the correct table owner.

Since Iceberg 1.1.0 there is now a way to provide a table owner via a
table property so we can make the create table operation to take one
step again.
https://github.com/apache/iceberg/pull/5763
https://github.com/apache/iceberg/pull/6154

This patch implements this behavior, when creating an Iceberg table
through HiveCatalog, specifying HMS_TABLE_OWNER as
hive.metastore.table.owner in properties can do it in one step.

Testing:
- Use the existing test test_iceberg.py test_table_owner.

Change-Id: I56ef7929449105af571d1fb9cb585d9b0733a39d
Reviewed-on: http://gerrit.cloudera.org:8080/20646
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>