IMPALA-12545: Use 'unique_database' in test_restart_services.py::TestRestart
Some tests in 'test_restart_services.py::TestRestart' create tables but
don't use a 'unique_database' for it. The table and column names are
sometimes the same. The tests are run serially but if a table is not
deleted successfully it may interfere with other tests.

This change introduces 'unique_database' in tests that create tables.
After this, explicitly deleting the tables is no longer necessary as the
framework takes care of it.

Testing:
 - all affected tests pass

Change-Id: Id4c2cfb669d5ff9e4d8110a0035bae1147e2db5a
Reviewed-on: http://gerrit.cloudera.org:8080/20662
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>