IMPALA-3268: Add support for SHOW VIEWS statement
This patch adds the support for the SHOW VIEWS statement and does not
change the current behavior of the SHOW TABLES statement. It is
recommended that we start the catalog server with
'--pull_table_types_and_comments=true' when executing this command
so that the table type of every Table is always available.

Specifically, after this patch, a user will be able to retrieve from
Impala the names of the Table's whose getTableType() evaluate to
TImpalaTableType.VIEW that match an optional pattern.

Similar to the SHOW TABLES statement, the following are supported
after this patch.
1. SHOW VIEWS.
2. SHOW VIEWS "<P1>|<P2>|<P3>", where <PN> denotes the N-th pattern,
   "|" denotes choice and each pattern may contain wildcards denoted by
   "*".
3. SHOW VIEWS IN <db>, where <db> denotes the name of a database.
Refer to the added end-to-end test cases for further details.

Recall that Hive distinguishes materialized views from views. To align
Impala's behavior with that of Hive with respect to the SHOW VIEWS
statement, in this patch we introduce
TImpalaTableType.MATERIALIZED_VIEW as a table type that corresponds to
the table type of 'MATERIALIZED_VIEW' for a
org.apache.hadoop.hive.metastore.api.Table. Due to this we have to
revise the mappings from Hive's table types to Impala's table types in
various places and have to slightly modify the logic used to
instantiate a table with an appropriate type.

Testing:
 - Added various frontend and end-to-end tests to verify the behavior
   of the SHOW VIEWS statement.

Change-Id: I321fc5350392a815949a4e7d2a64d60466689788
Reviewed-on: http://gerrit.cloudera.org:8080/20669
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>