IMPALA-12513: Allow to reset metadata when the CatalogD becomes active
When switching active catalogd, the loaded metadata in the standby
catalogd may not be the latest. A backend flag should be provided
to control whether to reset metadata when the catalogd become active.
Adds the following startup flags for catalogd:
'catalogd_ha_reset_metadata_on_failover'. Default is false. If true,
reset all metadata when the catalogd becomes active.

Testing:
- Added a test case to start both catalogds with flag
   'catalogd_ha_reset_metadata_on_failover' as true.
- Passed core tests

Change-Id: I2b54f36f96e7901499105e51790d8e2300d7fea9
Reviewed-on: http://gerrit.cloudera.org:8080/20614
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>