IMPALA-11542: Implement pre-allocation in LLVM memory manager
Implements up-front allocation for the LLVM memory manager to avoid
disparate sections on ARM which can exceed the 4GB limit for ADRP
instructions and crash (or hit an assertion in debug mode).

This is an LLVM issue that we're fixing by providing a custom memory
manager. See the JIRA and
https://discourse.llvm.org/t/llvm-rtdyld-aarch64-abi-relocation-restrictions/74616
for further discussion.

Testing:
- passes debug test run on ARM
- pre-commit tests

Change-Id: I9f224edcdbdcb05fce663c18b4a5f03c8e985675
Reviewed-on: http://gerrit.cloudera.org:8080/20692
Tested-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>