IMPALA-12555: Point Maven cache downloader to current location
bin/jenkins/populate_m2_directory.py is used during bootstrap_system.sh
to prime the local .m2 cache for Maven. This preloads the majority of
common Java dependencies for faster front-end builds.

The priming bundle is generated during nightly builds of the
'all-build-options' job running on the master branch. The downloader
script then reaches out to jenkins.impala.io to locate and download the
generated tarball.

This download has been failing for the past few weeks for a banal
reason: all the jobs for the upstream precommit environment were
migrated from Ubuntu 16.04 to Ubuntu 20.04, which was also reflected in
the job names. However, bin/jenkins/populate_m2_directory.py never
received the update to point it to the current version of the
all-build-options job, and the usable builds in the old location have
all aged out of Jenkins.

This patch points the job to the right location to restore cache
priming.

Change-Id: Id494fb7f24f1364a96526b440c8a0c4b6feda588
Reviewed-on: http://gerrit.cloudera.org:8080/20698
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>