IMPALA-12557: DELETE throws DateTimeParseException when deleting from time-partitioned table
There's a bug in IcebergDeleteSink that prevents Impala from
successfully executing a DELETE operation on Iceberg tables. During
DELETE we retrieve partition values from the virtual column
ICEBERG__PARTITION__SERIALIZED. This contains the transformed values,
e.g. in case of DAY-partitioning it contains the number of days since
the UNIX epoch.

Currently IcebergDeleteSink just uses these values as they are.
There are two problems with this. First, we want to place the delete
files under human-readable partition directories like other engines
do, and like our own INSERT statement does. I.e. we want a partition
directory /ts_day=2023-11-11/ and not /ts_day=19672/. The other problem
is that 'IcebergUtil.partitionDataFromDataFile()' also expects the
human-readable representation. This could be resolved at the CatalogD
side to just accept the integer values, but then we would still need
some logic in the IcebergDeleteSink to generate the human-readable
values for the file paths.

Moreover, partition values from INSERT statements are also
received in the human-readable representation at the Catalog.

This patch fixes the error by adding functions that transforms the
partition values to their human-readable representations. This is
done in the IcebergDeleteSink, so the Catalog-side logic is not
affected.

The above only affects the time-based transforms (YEAR, MONTH, DAY,
HOUR), as other partition transform values don't use different
representations.

Some notes on HOUR transform and daylight saving time:
There is no 1:1 mapping between an offset and the human-readable
representation in a timezone that has daylight saving time. This is not
an issue, as Impala's TIMESTAMP type is timezone-less. This also won't
be an issue for the TIMESTAMPTZ type as timestamp values are normalized
to UTC when stored, and UTC doesn't have daylight saving time.

Testing:
 * C++ backend tests
 * E2E tests for all time-based transforms, and also partition evolution
 * Also added an extra test about TRUNCATEing numeric values which was
   untested

Change-Id: I1cfeaed6409289663eb0f65b1ee2ecebd93e6118
Reviewed-on: http://gerrit.cloudera.org:8080/20711
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>