IMPALA-12577: Correctly update last-synced-event-time
There are two metrics tracking the progress of HMS event processing:
last-synced-event-id and last-synced-event-time. We have several places
that only update last-synced-event-id without updating
last-synced-event-time which is used to calculate the lag of event
processing. Users might set up alerts on the lag so we'd better keep it
updated.

Tests
 - Add code to verify the metrics in MetastoreEventsProcessorTest
 - Fix a potential NullPointer error in HiveJdbcClientPool.java
 - Ran CORE test

Change-Id: I033ace83221bc5e0c84c19aac78a41c1eeda31f2
Reviewed-on: http://gerrit.cloudera.org:8080/20732
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>