IMPALA-12001: Informative error message for complex types with DISTINCT
Before this change, queries with SELECT DISTINCT on a complex type
failed.

With structs, we got a FE exception:
  use functional_parquet;
  select distinct(struct_val) from alltypes_structs;

  ERROR: IllegalStateException: null

With collections, the BE hits a DCHECK and crashes:
  use functional_parquet;
  select distinct(arr1) from complextypes_arrays;

  Socket error 104: [Errno 104] Connection reset by peer

Aggregate functions with complex DISTINCT parameters also failed without
a clear error message. For example:
  select count(distinct struct_val) from alltypes_structs;
  select count(distinct arr1) from complextypes_arrays;

To support DISTINCT for complex types we would need to implement
equality and hash for them. We are not planning to do it in the near
future, so this change introduces informative error messages in these
cases.

Testing:
 - added test queries for SELECT DISTINCT and SELECT COUNT(DISTINCT ...)
   with arrays, maps and structs, expecting the correct error messages.

Change-Id: Ibe2642d1683a10fd05a95e2ad8470d16f0d5242c
Reviewed-on: http://gerrit.cloudera.org:8080/20752
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>