IMPALA-12590: Fix dmesg call during precommit for Ubuntu 20.04
Ubuntu 20.04 locked down access to the kernel messages, so a call to
'dmesg' can succeed only when executed with elevated privileges.

This could be a problem during Impala precommit runs, as the finalizer
script uses 'dmesg' to detect potential OOM-kills during the run.

This patch adds an "escalation" step to the dmesg call: if the regular
call fails, it issues a second call via 'sudo'.

Change-Id: Ic20193740c6e5cb9e8e155c03bede55184875de5
Reviewed-on: http://gerrit.cloudera.org:8080/20763
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>