IMPALA-12744: Support for regr_count() aggregate/analytic function
The regr_count() function can be used both as an aggregate and
an analytic function and is commonly used in regression analysis.

regr_count(y, x) returns an integer that is the number of non-null
number pairs. It indicates how many observations are included in the
analysis.

Testing:
The functions are extensively tested and cross-checked with Hive.
The tests can be found in aggregation.test.

Change-Id: I773d1e0edc8a9c8ee003f75721f4844685b2eb38
Reviewed-on: http://gerrit.cloudera.org:8080/20936
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>